Refactor of form management
