+++ src/EditModel/event-program/create-data-entry-form/CustomForm.component.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import EditCustomFormProgramStage from '../data-entry-form/EditCustomFormProgramStage';
 
-export default function CustomForm(props) {
-    return <EditCustomFormProgramStage {...props} />;
-}
+const CustomForm = props => (
+    <EditCustomFormProgramStage {...props} />
+);
+
+export default CustomForm;
