+++ src/EditModel/event-program/create-data-entry-form/SortableSectionDataList.component.js
@@ -2,7 +2,7 @@ import React, { PropTypes } from 'react';
 import { SortableContainer, SortableElement } from 'react-sortable-hoc';
 import DragHandle from './DragHandle.component';
 import { grey100 } from 'material-ui/styles/colors';
-import { ActionButton } from './SectionForm.component';
+import ActionButton from './ActionButton.component';
 
 const styles = {
     dataElement: {
@@ -27,37 +27,23 @@ const styles = {
     },
 };
 
-const SortableSectionDataList = SortableContainer(({ sectionDataElements, onDataElementRemoved }) => <div>
-    { sectionDataElements.map((dataElement, index) => (
-        <SortableSectionDataElement
-            dataElement={dataElement}
-            onRemove={() => { onDataElementRemoved(dataElement.id); }}
-            index={index}
-            key={`item-${index}`}
-            sortIndex={index}
-        />
-        ))}
-</div>);
+const SectionDataElement = ({ first, dataElement, onRemove }) => {
+    const divStyle = {
+        ...styles.dataElement,
+        marginTop: first ? '0px' : '4px',
+    };
 
-const SortableSectionDataElement = SortableElement(({ index, sortIndex, dataElement, onRemove }) => (
-    <SectionDataElement first={sortIndex === 0} index={index} sortOrder={sortIndex} dataElement={dataElement} onRemove={onRemove} />
-));
-
-const SectionDataElement = ({ first, sortOrder, dataElement, onRemove }) => (
-    <div
-        style={{
-            ...styles.dataElement,
-            marginTop: first ? '0px' : '4px',
-        }}
-    >
-        <div style={styles.row}>
-            <DragHandle />
-            <div style={styles.horizontalSpace} />
-            {dataElement.displayName}
+    return (
+        <div style={divStyle}>
+            <div style={styles.row}>
+                <DragHandle />
+                <div style={styles.horizontalSpace} />
+                {dataElement.displayName}
+            </div>
+            <ActionButton onClick={onRemove} icon="clear" />
         </div>
-        <ActionButton onClick={onRemove} icon="clear" />
-    </div>
-);
+    );
+};
 
 SectionDataElement.propTypes = {
     dataElement: PropTypes.shape({
@@ -65,5 +51,18 @@ SectionDataElement.propTypes = {
         displayName: PropTypes.string.isRequired,
     }),
 };
+
+const SortableSectionDataElement = SortableElement(SectionDataElement);
+const SortableSectionDataList = SortableContainer(({ sectionDataElements, onDataElementRemoved }) => <div>
+    { sectionDataElements.map((dataElement, index) => (
+        <SortableSectionDataElement
+            first={index === 0}
+            dataElement={dataElement}
+            onRemove={() => { onDataElementRemoved(dataElement.id); }}
+            index={index}
+            key={`item-${index}`}
+        />
+        ))}
+</div>);
 
 export default SortableSectionDataList;
